Changed the initial DesignEdit value to true so the editor is immediately in editable mode.
Deleted Resources/BasicTestFileX.html. This file does not belong in the sample,
as it is intended as part of the in the Testing sample to provide a convenient method
of developing and testing underlying javascript to be injected into WebView2.

The Resources/LargeTestDocument.html and Resources/SmallTestFile.html files were
modified to restore the original styling of the tables.